+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file.
 - Made conversation history of threads be resumable after Cyrus restarts
 - Fixed the issue with continuity of conversation in a thread, after the first comment
 
+### Changed
+- Renamed `.edge-config.json` to `.cyrus-edge-config.json`
+
 ## [0.1.32] - 2025-01-09
 
 ### CLI